feat: remove animations and move navigators outside components
