+++ sites/tutors-course-reader/src/App.svelte
@@ -12,20 +12,20 @@
   import Wall from "./pages/Wall.svelte";
   import Lab from "./pages/Lab.svelte";
   import NotFound from "./pages/support/NotFound.svelte";
-  import NavBar from "./components/navigators/NavBar.svelte";
-  import PageHeader from "./components/navigators/PageHeader.svelte";
-  import Footer from "./components/navigators/footer/Footer.svelte";
-  import BackToTop from "./components/navigators/BackToTop.svelte";
+  import NavBar from "./navigators/NavBar.svelte";
+  import PageHeader from "./navigators/PageHeader.svelte";
+  import Footer from "./navigators/footer/Footer.svelte";
+  import BackToTop from "./navigators/BackToTop.svelte";
   import Logout from "./pages/support/Logout.svelte";
   import TutorsTerms from "./pages/support/TutorsTerms.svelte";
   import { CourseService } from "./reader-lib/services/course-service";
   import { handleAuthentication } from "./reader-lib/services/auth-service";
   import { AnalyticsService } from "./reader-lib/services/analytics-service";
   import Search from "./pages/Search.svelte";
-  import InfoBar from "./components/navigators/sidebars/InfoBar.svelte";
-  import CalendarBar from "./components/navigators/sidebars/CalendarBar.svelte";
-  import OnlineBar from "./components/navigators/sidebars/OnlineBar.svelte";
-  import TocBar from "./components/navigators/sidebars/TocBar.svelte";
+  import InfoBar from "./navigators/sidebars/InfoBar.svelte";
+  import CalendarBar from "./navigators/sidebars/CalendarBar.svelte";
+  import OnlineBar from "./navigators/sidebars/OnlineBar.svelte";
+  import TocBar from "./navigators/sidebars/TocBar.svelte";
   import Note from "./pages/Note.svelte";
   import { MetricsService } from "./reader-lib/services/metrics-service";
   import { currentLo, infoDrawer, calendarDrawer, onlineDrawer, tocDrawer, storeTheme } from "./stores";
