+++ sites/tutors-course-reader/src/navigators/CourseNavigator.svelte
