+++ sites/tutors-course-reader/src/navigators/PageHeader.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import { currentCourse, portfolio } from "../../stores";
+  import { currentCourse, portfolio } from "../stores";
   import { Breadcrumbs } from "tutors-ui";
   import IconBar from "tutors-reader-lib/src/iconography/IconBar.svelte";
 </script>
