+++ sites/tutors-course-reader/src/navigators/footer/Footer.svelte
