+++ sites/tutors-course-reader/src/navigators/sidebars/TocBar.svelte
@@ -1,24 +1,22 @@
 <script lang="ts">
   import { beforeUpdate, getContext } from "svelte";
   import type { Course } from "tutors-reader-lib/src/models/course";
-  import type { CourseService } from "../../../reader-lib/services/course-service";
-  import { infoDrawer } from "../../../stores";
-  import { convertMd } from "tutors-reader-lib/src/utils/markdown-utils";
+  import type { CourseService } from "../../reader-lib/services/course-service";
+  import CourseNavigator from "../CourseNavigator.svelte";
+  import { tocDrawer } from "../../stores";
 
   let course: Course = null;
   const cache: CourseService = getContext("cache");
-  let courseInfo = "";
+
   let display = false;
   beforeUpdate(() => {
     course = cache.course;
     if (course) {
       display = true;
-      courseInfo = convertMd(course.lo.contentMd, null);
     }
   });
-
   const drawerClose: any = () => {
-    infoDrawer.set(false);
+    tocDrawer.set(false);
   };
 </script>
 
@@ -26,7 +24,5 @@
   <button class="btn btn-icon bg-primary-500 text-white" on:click={drawerClose}><span class="font-bold">X</span></button>
 </div>
 <div class="px-12 py-4">
-  <prose class="prose dark:prose-invert">
-    {@html courseInfo}
-  </prose>
+  <CourseNavigator {course} />
 </div>
